Deleted unused components and renamed components for better clarity
