+++ src/components/LandingPageContainer.js
@@ -2,12 +2,12 @@ import React, { useState, useEffect } from "react";
 import axios from "axios";
 import { useMediaQuery } from "react-responsive";
 
-import HeaderImage from "./HeaderImage";
-import HeaderInfo from "./HeaderInfo";
+import HeaderInfo from "./LandingPageInfo";
 import LandingPageCard from "./LandingPageCard";
 import LoadingSpinner from "./assets/LoadingSpinner";
 import ErrorMessage from "./assets/ErrorMessage";
 import { imageData } from "./assets/locationData";
+import headerImage from "../img/HeaderImage.webp";
 
 const LandingPageContainer = () => {
   // ---------------- States ------------------- //
@@ -27,9 +27,7 @@ const LandingPageContainer = () => {
       setIsLoading(true);
       setIsError(false);
       try {
-        const response = await axios.get(
-          `https://backend-zaalverhuur.herokuapp.com/api`
-        );
+        const response = await axios.get(`https://backend-zaalverhuur.herokuapp.com/api`);
         console.log("API response Landing Page", response.data);
         setApiData(response.data);
       } catch (error) {
@@ -50,10 +48,19 @@ const LandingPageContainer = () => {
     minHeight: "20vh",
   };
 
+  const imageStyle = {
+    maxWidth: "1300px",
+    width: "100%",
+    maxHeight: "200px",
+    marginBottom: "15px",
+    objectFit: "cover",
+    display: isMobile ? "none" : "block",
+  };
+
   // ---------------- Render ------------------- //
   return (
     <div>
-      {!isMobile && <HeaderImage />}
+      <img alt="headerimage" style={imageStyle} src={headerImage} />
       <HeaderInfo />
 
       <div style={divStyle}>
@@ -63,7 +70,6 @@ const LandingPageContainer = () => {
         {apiData.map((mappedApiData, index) => {
           return (
             // ---- Temporary rudementary filters. TODO: Move to back-end ---- //
-            mappedApiData.locationID <= 1371 &&
             mappedApiData.locationID !== 1366 && (
               <LandingPageCard
                 link={mappedApiData.locationID}
